Rename iterators package to access
The purpose of the classes in the package are to
provide access to different parts of the score and
not just provide iterators.
