+++ src/main/java/org/wmn4j/notation/Part.java
@@ -3,7 +3,7 @@
  */
 package org.wmn4j.notation;
 
-import org.wmn4j.notation.iterators.PartIterator;
+import org.wmn4j.notation.access.PartIterator;
 
 import java.util.Iterator;
 import java.util.List;
