+++ src/main/java/org/wmn4j/notation/access/PartWiseScoreIterator.java
@@ -1,7 +1,7 @@
 /*
  * Distributed under the MIT license (see LICENSE.txt or https://opensource.org/licenses/MIT).
  */
-package org.wmn4j.notation.iterators;
+package org.wmn4j.notation.access;
 
 import org.wmn4j.notation.Durational;
 import org.wmn4j.notation.Part;
