+++ src/main/java/org/wmn4j/notation/access/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides different methods, such as iterators, for accessing the notation elements in a score.
+ */
+package org.wmn4j.notation.access;
