--- src/main/java/org/wmn4j/notation/iterators/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Contains iterators for iterating through the elements of notation that
- * contain other notation elements.
- */
-package org.wmn4j.notation.iterators;
